requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
bcrypt>=4.0.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
sqlite3